+++ pages/home.tsx
@@ -1,0 +1,12 @@
+import Header from "components/header";
+
+const Homepage = () => {
+  return (
+    <div>
+      <Header />
+      <h1>Homepage</h1>
+    </div>
+  );
+};
+
+export default Homepage;
