+++ pages/index.module.scss
@@ -1,27 +1,32 @@
 @import "styles/common";
 
-.homepage {
-  position: relative;
-  height: 100%;
+.wrapper {
+  height: 100vh;
+  width: 100vw;
 
-  .home {
-    width: 100%;
-    @extend .alignCen;
+  .homepage {
+    position: relative;
     height: 100%;
-    padding: 1rem;
-    padding-top: 5rem;
-    overflow: scroll;
 
-    &::-webkit-scrollbar {
-      display: none;
-    }
-
-    .row {
-      @extend .flexCenter;
-      gap: 1rem;
+    .home {
       width: 100%;
+      @extend .alignCen;
       height: 100%;
-      margin-left: 100px;
+      padding: 1rem;
+      padding-top: 5rem;
+      overflow: scroll;
+
+      &::-webkit-scrollbar {
+        display: none;
+      }
+
+      .row {
+        @extend .flexCenter;
+        gap: 1rem;
+        width: 100%;
+        height: 100%;
+        margin-left: 100px;
+      }
     }
   }
 }
