+++ src/components/customizer/Customizer.tsx
@@ -3,6 +3,7 @@ import Image from "next/image";
 
 import { modules, innereyes, outereyes, logos } from "lib/data";
 import Dropdown from "components/dropdown";
+import RadioInput from "components/radio-inputs";
 
 import stl from "./Customizer.module.scss";
 
@@ -102,36 +103,7 @@ const Customizer = ({ setStyles }: Props) => {
             })
           }
         />
-        <div className={stl.radioInputs}>
-          <label className={stl.radio} onClick={() => setFormat("jpg")}>
-            <input
-              type="radio"
-              name="format"
-              checked={format === "jpg"}
-              onChange={(e) => console.log(e.target.value)}
-            />
-            <span className={stl.label}>JPG</span>
-          </label>
-          <label className={stl.radio} onClick={() => setFormat("png")}>
-            <input
-              type="radio"
-              name="format"
-              checked={format === "png"}
-              onChange={(e) => console.log(e.target.value)}
-            />
-            <span className={stl.label}>PNG</span>
-          </label>
-
-          <label className={stl.radio} onClick={() => setFormat("svg")}>
-            <input
-              type="radio"
-              name="format"
-              checked={format === "svg"}
-              onChange={(e) => console.log(e.target.value)}
-            />
-            <span className={stl.label}>SVG</span>
-          </label>
-        </div>
+        <RadioInput format={format} setFormat={setFormat} />
       </div>
     </div>
   );
