+++ src/components/radio-inputs/RadioInputs.module.scss
@@ -1,0 +1,40 @@
+@import "styles/common";
+
+.radioInputs {
+  position: relative;
+  display: flex;
+  flex-wrap: wrap;
+  margin: 12px auto 0 auto;
+  border-radius: 0.5rem;
+  background-color: #eee;
+  box-sizing: border-box;
+  box-shadow: 0 0 0px 1px rgba(0, 0, 0, 0.06);
+  padding: 0.25rem;
+  width: 300px;
+  font-size: 1rem;
+
+  .radio {
+    flex: 1 1 auto;
+    text-align: center;
+
+    input {
+      display: none;
+    }
+
+    .label {
+      @extend .flexCenter;
+      border-radius: 0.5rem;
+      border: none;
+      padding: 0.5rem 0;
+      color: #334155;
+      transition: all 0.15s ease-in-out;
+      cursor: pointer;
+    }
+
+    input:checked + .label {
+      background: #1e90ff;
+      font-weight: 600;
+      color: #fff;
+    }
+  }
+}
