+++ src/components/radio-inputs/index.tsx
@@ -1,0 +1,3 @@
+import RadioInput from "./RadioInputs";
+
+export default RadioInput;
